refactor(test\...\controller) : change position of WinResultManager Test
